requirements.txt
flask
pylance
flask-sqlalchemy
mysql-connector-python
python-dotenv
passlib
Flask-Migrate
pymysql
pytest
boto3
cryptography
